requirements.txt
aiohttp==3.6.2
aiojobs==0.2.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
gunicorn==20.0.4
idna==2.8
multidict==4.7.4
python-dotenv==0.12.0
requests==2.22.0
urllib3==1.25.8
yarl==1.4.2
